chore: 기존 IconWithText를 레거시로 변경
